Separate out block getter from RPC serialization
This adds a new method to the block chain by adding a new matching
method getBlockAtSequence which matches getHeaderAtSequence. It also
does not conflate data fetching with serialization so getBlocks is a lot
cleaner now.
